luxe_rma/requirements.txt
Flask
Flask-Mail
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
Werkzeug
psycopg2-binary
PyJWT
pytest
